move decorations out of cell renderer
